moved reducers into redux directory
